Handle fragment_container in MainActivity.kt

+++ app/src/main/java/com/example/recipebox/RecipeDetailFragment.kt
@@ -20,8 +20,8 @@ class RecipeDetailFragment : Fragment() {
         return inflater.inflate(R.layout.fragment_recipe_detail, container, false)
     }
 
-    fun setRecipe(id: Int) {
-        this.recipeId = id;
+    fun setRecipe(id: Long) {
+        this.recipeId = id.toInt();
     }
 
     override fun onStart() {
